fix(query): 修 time ± duration 的纳秒截断（TCK 回归），并对齐 duration 的比较语义
一、真回归：TimeValue::nanos 被收窄成 int8
`TimeValue addDuration()` 末尾写成 `result.nanos = static_cast<int8_t>(nanos)`（c98867e 打包时引入，
nanos 是 int32、值域到 999'999'999），999'999'999 截断成 -1，于是 `time/localtime ± duration`
的亚秒部分出错。TCK 实测：`localtime('12:31:14.000000001') - duration({...seconds:70, nanoseconds:2})`
应得 `20:18:03.999999999`，当时给的是 `20:18:03.000000001`。
这条解释了 CI TCK 报告里 4 条 Temporal8 [2]/[3] 的回归；改回 int32 后
`Temporal8.feature` 本地 27/27 通过（此前 23/27）。顺带把同一处的
`result.hour = static_cast<int32_t>(hour)` 改成 int8（与实际字段一致）。

二、duration 的排序比较对齐 neo4j（B4）
neo4j 里 duration 只能做 = / <>：`<` / `>` 返回 null；`<=` / `>=` 等价于 `(a < b) OR (a = b)`，
a < b 是 null，所以只有相等时给 true、不等时也是 null。四个比较内核按此实现
（此前返回布尔值）。ORDER BY / min / max 仍走 compareDuration/durationOrderNanos，不受影响。

三、补单测（同时回应 Codecov 的 diff 覆盖率）
* TemporalTimeSubsecondArithmeticKeepsNanos：上面的回归（期望值取 Temporal8 的 Examples）；
* TemporalDurationOrderingOperatorsAreNullLikeNeo4j：三值逻辑（true/false/null）+ WHERE 过滤 + min/max；
* ScalarFunctionsOnEntityAndCollectionValues：值打包后"非 typed-batch"的通用分支
  （path/graph 函数、列表函数与下标、转换函数的 TypeError 分支）——这些正是 diff 覆盖率为 0% 的文件。

验证：`query_executor_tests` 三个新用例通过；`Temporal8.feature` 27/27；`check-format.sh` 通过。
遗留：CI TCK 报告里的另外 ~23 条回归（list 的 ORDER BY、min/max over list/mixed、跨类型比较）
与本条同源（打包引入），尚未定位到具体代码，下一步继续查。

+++ src/common/types/temporal_value.cpp
@@ -478,10 +478,12 @@ TimeValue addDuration(const TimeValue& temporal, const DurationValue& duration) 
         hour -= bh;
     }
 
-    result.hour = static_cast<int32_t>(hour);
+    result.hour = static_cast<int8_t>(hour);
     result.minute = static_cast<int8_t>(minute);
     result.second = static_cast<int8_t>(second);
-    result.nanos = static_cast<int8_t>(nanos);
+    // nanos 是 int32（0..999'999'999）：收窄成 int8 会把 999'999'999 截成 -1，
+    // 于是 time / localtime ± duration 的亚秒部分出错（TCK Temporal8 [2]/[3] 回归）。
+    result.nanos = static_cast<int32_t>(nanos);
     return result;
 }
